% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Present output from a dataset classified by \code{MLWIC}}
\usage{
make_output(output_location, model_dir, output_name = "output.csv",
  saved_predictions = "model_predictions.txt")
}
\arguments{
\item{output_location}{Absolute path where you want the output csv stored. This path
must exist on your computer.}

\item{model_dir}{Absolute path to the location where you stored the L1 folder
that you downloaded from github.}

\item{output_name}{Desired name of the output file}

\item{saved_predictions}{This is the file name where you stored predictions when you ran
\code{classify}. If you used the default in that function, you can use the default here.}
}
\description{
\code{make_output} will make a clean csv presenting the results from your run
of \code{classify}.
}
