% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLWIC_eval.R
\name{MLWIC_eval}
\alias{MLWIC_eval}
\title{Evaluate images using the trained model}
\usage{
MLWIC_eval(path_prefix, data_info, model_dir,
  save_predictions = "model_predictions.txt",
  python_loc = "/anaconda2/bin/", num_classes = 28, delimiter = ",",
  log_dir = "USDA182")
}
\arguments{
\item{path_prefix}{Absolute path to location of the images on your computer (or computing cluster).
All images must be stored in one folder}

\item{data_info}{Name of a csv containing the file names of each image.
This file must have only two columns and NO HEADERS. The first column must be the file name of the image
The second column can be the number corresponding to the species or group in the image.
See Table 1 in Tabak et al. for the numbers (if using the built in model. If you do not know the species in the image,
put a zero in each row of column 2.}

\item{model_dir}{Absolute path to the location where you stored the L1 folder
that you downloaded from github.}

\item{save_predictions}{File name where model predictions will be stored.
You should not need to change this parameter.
After running this function, you will run \code{MLWIC_make_output} to
make the output in a more viewer friendly format}

\item{python_loc}{The location of python 2.7 on your machine. If you are
using a Macintosh, the default is the likely location}

\item{num_classes}{The number of classes in your model. If you are using
the Species Level model from Tabak et al., the number is `28`.}

\item{delimiter}{this will be a `,` for a csv.}

\item{log_dir}{If you trained a model with \code{MLWIC_train}, this
will be the log_directory that you specified when using that function.
If you are using the built in model, the default is appropriate.}
}
\description{
\code{MLWIC_eval} Uses the Species Level model from Tabak et al. (the built in model) to predict
the species in each image. If you trained a model using \code{MLWIC_train},
this function can also be used to evalute images using the model developed by
\code{MLWIC_train} by specifying the \code{log_dir} of the trained model.
}
