% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLWIC_setup.R
\name{MLWIC_setup}
\alias{MLWIC_setup}
\title{Setup your computer to run \code{MLWIC}}
\usage{
MLWIC_setup(python_loc = "/anaconda2/bin/python2.7")
}
\arguments{
\item{python_loc}{The location of python 2.7 on your machine. If you are
using a Macintosh, the default is the likely location
@export}
}
\description{
\code{MLWIC_setup} installs necessary R packages on your computer. You will
need to run this before running \code{MLWIC_eval} and \code{MLWIC_train}.
}
