% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup your computer to run \code{MLWIC}}
\usage{
setup(python_loc = "/anaconda2/bin/python")
}
\arguments{
\item{python_loc}{The location of python 2.7 on your machine. If you are
using a Macintosh, the default is the likely location.}
}
\description{
\code{setup} installs necessary R packages on your computer. You will
need to run this before running \code{classify} and \code{train}.
}
